Move a bunch of function from domain.py to app.py because it's much simpler to have them here
